requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.3
charset==1.0.1
click==6.7
colorama==0.3.9
constantly==15.1.0
cryptography==2.0.3
cssselect==1.0.1
DBUtils==1.2
Django==1.11.2
django-bootstrap-toolkit==2.15.0
docopt==0.6.2
Flask==0.12.2
hyperlink==17.3.1
idna==2.5
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.9.6
lxml==3.7.3
MarkupSafe==1.0
mod-wsgi==4.5.17+ap24vc14
mysqlclient==1.3.10
olefile==0.44
parsel==1.2.0
Pillow==4.2.1
prettytable==0.7.2
pyasn1==0.3.2
pyasn1-modules==0.0.11
pycparser==2.18
PyDispatcher==2.0.5
PyMySQL3==0.5
pyOpenSSL==17.2.0
pypiwin32==220
pytz==2017.2
queuelib==1.4.2
redis==2.10.6
requests==2.17.3
Scrapy==1.4.0
selenium==3.4.3
service-identity==17.0.0
six==1.10.0
tabulate==0.7.7
telnetlib3==1.0.0
Twisted==17.5.0
urllib3==1.21.1
w3lib==1.18.0
Werkzeug==0.12.2
xlrd==1.0.0
zope.interface==4.4.2
